+++ crates/kira/src/track/effect/eq_filter/builder.rs
@@ -8,17 +8,17 @@ const COMMAND_CAPACITY: usize = 8;
 
 pub struct EqFilterBuilder {
 	pub kind: EqFilterKind,
-	pub frequency: Value<f32>,
-	pub gain: Value<f32>,
-	pub q: Value<f32>,
+	pub frequency: Value<f64>,
+	pub gain: Value<f64>,
+	pub q: Value<f64>,
 }
 
 impl EqFilterBuilder {
 	pub fn new(
 		kind: EqFilterKind,
-		frequency: impl Into<Value<f32>>,
-		gain: impl Into<Value<f32>>,
-		q: impl Into<Value<f32>>,
+		frequency: impl Into<Value<f64>>,
+		gain: impl Into<Value<f64>>,
+		q: impl Into<Value<f64>>,
 	) -> Self {
 		Self {
 			kind,
