+++ crates/kira/src/track/effect/eq_filter/handle.rs
@@ -12,7 +12,7 @@ impl EqFilterHandle {
 	/// Sets the frequency of the filter (in hertz).
 	pub fn set_frequency(
 		&mut self,
-		frequency: impl Into<Value<f32>>,
+		frequency: impl Into<Value<f64>>,
 		tween: Tween,
 	) -> Result<(), CommandError> {
 		self.command_producer
@@ -23,7 +23,7 @@ impl EqFilterHandle {
 	/// Sets the gain of the filter (in hertz).
 	pub fn set_gain(
 		&mut self,
-		gain: impl Into<Value<f32>>,
+		gain: impl Into<Value<f64>>,
 		tween: Tween,
 	) -> Result<(), CommandError> {
 		self.command_producer
@@ -32,7 +32,7 @@ impl EqFilterHandle {
 	}
 
 	/// Sets the q value of the filter (in hertz).
-	pub fn set_q(&mut self, q: impl Into<Value<f32>>, tween: Tween) -> Result<(), CommandError> {
+	pub fn set_q(&mut self, q: impl Into<Value<f64>>, tween: Tween) -> Result<(), CommandError> {
 		self.command_producer
 			.push(Command::SetQ(q.into(), tween))
 			.map_err(|_| CommandError::CommandQueueFull)
